IRISHUB-358: refactor version command for iris, iriscli, irislcd
