+++ client/upgrade/cli/query.go
@@ -7,6 +7,7 @@ import (
 	"github.com/cosmos/cosmos-sdk/wire"
 	authcmd "github.com/cosmos/cosmos-sdk/x/auth/client/cli"
 	"github.com/irisnet/irishub/modules/upgrade"
+	irisVersion "github.com/irisnet/irishub/version"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 	"os"
@@ -90,5 +91,41 @@ func GetCmdQuerySwitch(storeName string, cdc *wire.Codec) *cobra.Command {
 	cmd.Flags().String(flagProposalID, "", "proposalID of upgrade swtich being queried")
 	cmd.Flags().String(flagVoter, "", "Address sign the switch msg")
 
+	return cmd
+}
+
+func GetCmdVersion(storeName string, cdc *wire.Codec) *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "node-version",
+		Short: "Show full node version info",
+		RunE: func(cmd *cobra.Command, args []string) error {
+
+			fmt.Printf("v%s\n", irisVersion.Version)
+
+			ctx := context.NewCLIContext()
+
+			var res_versionID []byte
+			var err error
+			res_versionID, err = ctx.QueryStore(upgrade.GetCurrentVersionKey(), storeName)
+			if err != nil {
+				return nil
+			}
+			var versionID int64
+			cdc.MustUnmarshalBinary(res_versionID, &versionID)
+
+			var res_version []byte
+			res_version, err = ctx.QueryStore(upgrade.GetVersionIDKey(versionID), storeName)
+			if err != nil {
+				return nil
+			}
+			var version upgrade.Version
+			cdc.MustUnmarshalBinary(res_version, &version)
+
+			fmt.Println(version.Id)
+			fmt.Println("Current version: Start Height    = ", version.Start)
+			fmt.Println("Current version: Proposal Id     = ", version.ProposalID)
+			return nil
+		},
+	}
 	return cmd
 }
