+++ cmd/iris/main.go
@@ -10,15 +10,15 @@ import (
 	"github.com/cosmos/cosmos-sdk/server"
 	"github.com/irisnet/irishub/app"
 	bam "github.com/irisnet/irishub/baseapp"
-	"github.com/irisnet/irishub/version"
 
+	"github.com/irisnet/irishub/tools/prometheus"
+	"github.com/irisnet/irishub/version"
 	"github.com/spf13/viper"
 	abci "github.com/tendermint/tendermint/abci/types"
 	"github.com/tendermint/tendermint/libs/cli"
 	dbm "github.com/tendermint/tendermint/libs/db"
 	"github.com/tendermint/tendermint/libs/log"
 	tmtypes "github.com/tendermint/tendermint/types"
-	"github.com/irisnet/irishub/tools/prometheus"
 )
 
 func main() {
@@ -57,9 +57,8 @@ func main() {
 	)
 
 	rootCmd.AddCommand(
-		client.GetCommands(
-			version.GetCmdVersion("upgrade", cdc),
-		)...)
+		version.ServeVersionCommand(cdc),
+	)
 
 	rootCmd.AddCommand(prometheus.MonitorCommand(cdc))
