+++ cmd/irislcd/main.go
@@ -1,10 +1,12 @@
 package main
 
 import (
-	_ "github.com/irisnet/irishub/client/lcd/statik"
-	"github.com/spf13/cobra"
 	"github.com/irisnet/irishub/app"
 	"github.com/irisnet/irishub/client/lcd"
+	_ "github.com/irisnet/irishub/client/lcd/statik"
+	"github.com/irisnet/irishub/version"
+	"github.com/spf13/cobra"
+	"github.com/tendermint/tendermint/libs/cli"
 )
 
 // rootCmd is the entry point for this binary
@@ -20,7 +22,15 @@ func main() {
 	cdc := app.MakeCodec()
 
 	rootCmd.AddCommand(
-		lcd.ServeLCDCommand(cdc),
+		lcd.ServeLCDStartCommand(cdc),
+		version.ServeVersionCommand(cdc),
 	)
-}
 
+	// prepare and add flags
+	executor := cli.PrepareMainCmd(rootCmd, "IRISLCD", app.DefaultCLIHome)
+	err := executor.Execute()
+	if err != nil {
+		// handle with #870
+		panic(err)
+	}
+}
